first task api

+++ java_domination_339-java_labs_2024/src/main/java/ru/ssau/tk/java_domination_339/java_labs_2024/functions/AbstractTabulatedFunction.java
@@ -76,7 +76,7 @@ public abstract class AbstractTabulatedFunction implements TabulatedFunction, Se
     }
 
     @Override
-    public int HashName() {
+    public long HashName() {
         return this.Name().hashCode();
     }
 }
